pyproject.toml
[tool.poetry]
name = "pyotr"
version = "0.1.0"
description = "OpenAPI-to-REST server and client frameworks for Python."
authors = ["Berislav Lopac <berislav@lopac.net>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
openapi-core = "^0.11.0"
pyyaml = "^5.1"
starlette = "^0.12.1"
uvicorn = "^0.8.2"
stringcase = "^1.2"
http3 = "^0.6.3"
requests = "^2.22"
pytest-cov = "^2.7"
typing-extensions = "^3.7"

[tool.poetry.dev-dependencies]
pytest = "^4.6"
pytest-flake8 = "^1.0"
pytest-asyncio = "^0.10.0"
pytest-mypy = "^0.3.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
